test: migrate saveFlow tests to orchestra, keep shard facade test in CLI
10 direct saveFlow tests now live in
maestro-orchestra/src/test/.../TestOutputWriterTest.kt and exercise
TestOutputWriter.saveFlow directly with filenamePrefix / logPrefix args.
CLI retains one test proving TestDebugReporter.saveFlow translates
shardIndex correctly into the shard-N+1 filename prefix.
